+++ client/src/components/Type/User.tsx
@@ -1,3 +1,5 @@
+import { TeacherType } from 'configs/List/config';
+
 export type User = {
   name: string;
   email: string;
@@ -24,4 +26,9 @@ export type UserRequest = {
   requestcategory: string[];
   note: string;
   date: string;
+};
+
+export type initialStateType = {
+  status: string;
+  value: TeacherType[];
 };
